+++ lib/falcon/limiter/socket.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2025, by Samuel Williams.
+# Copyright, 2025-2026, by Samuel Williams.
 
 module Falcon
 	module Limiter
@@ -25,7 +25,7 @@ module Falcon
 			ensure
 				if token = @token
 					@token = nil
-					token.release
+					token.close
 				end
 			end